servicos_software_p/api-visao/requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
pillow
requests